requirements.txt
Django==2.1.8
django-appconf==1.0.3
django-compressor==2.3
django-environ==0.4.5
django-fancybox==0.1.5
django-photologue==3.9
django-sass==0.1.1
django-sass-processor==0.7.3
django-sortedm2m==1.5.0
django-taggit==1.1.0
ExifRead==2.1.2
libsass==0.19.1
Pillow==6.0.0
pytz==2019.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
